Finish

+++ src/main/java/com/cbox/c1/entity/Event.java
@@ -7,7 +7,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 
 @Entity
 @Data
@@ -18,7 +17,7 @@ import java.time.LocalDateTime;
 public class Event {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer pEno;
+    private Long pEno;
     private String eventInfo;
     private LocalDate startDate;
     private LocalDate endDate;
